Fix cudahelper.check_cuda_available()
Because terralib.linklibrary doesn't consider LD_LIBRARY_PATH when resolving
recursive dependencies, we can no longer safely link libregent.so (linking will
fail if using the stub libcuda.so, which will not be in /usr/lib). The
cudahelper module would first load libregent.so, then search the set of loaded
symbols for cuInit (a libregent.so compiled with USE_CUDA=1 would automatically
pull in libcuda.so). This strategy will no longer work, so we instead always
link in libcuda.so (before libregent.so) if -fcuda is 1.