regent: Change calling convention to make order of region requirements stable independent of user's specification of constraints.
The calling convention for Regent now explicitly sorts region
requirements by the privilege/coherence/flags used:

  * Privileges: reads, writes, reduces (in that order)
  * Coherence: exclusive, atomic, simultaneous, relaxed (in that order)
  * Flags: sorted lexicographically on the name of the flag

This helps to ensure that the region requirements do not depend on
e.g. the order in which privileges are specific in the user's
program. It is also necessary to make a sane implementation of
Regent's calling convention that can match Regent's output without
having the exact input program text.