regent: Set Kepler compute capability to 3.2 to work around LLVM bug.
Technically, Kepler GPUs could be any of 3.0, 3.2, 3.5 or
3.7. However, LLVM's support for all compute capabilities other than
3.2 seems to be broken. Specifically, 3.0 and 3.5 use PTX version 3.2,
which is not new enough for certain sqrt instructions. I have been
unable to find a way to set the PTX version. Instead, we change the
Kepler compute capability to 3.2, which LLVM chooses to use PTX
version 4.0 for. This is a hack but seems to work well enough on
applications and Kepler GPUs we've tested.