runtime/legion: introduce a maximum replayable trace length
This does not affect the length of traces we may try to find in the
applications, but the size of traces that we insert into the trace
replayer objects. Traces that are larger than the maximum length
(intended to be reasonably large) will be broken up into sub-traces
that are smaller than the maximum length.