legion_prof_rs: Remove superfluous instance level code.
The code removed in this change appears to be related to a feature (or
planned feature?) of the legacy JavaScript viewer for viewing
instances at the point in time they were created, rather than ready
for use. While this might be a generally useful feature, it's not
clear that it ever worked. The current code is certainly not
sufficient to implement it, nor is it the way this change would be
implemented today. Since it costs space and time for something not
actually supported, this change removes it.